+++ tests/_async/test_storages.py
@@ -1,8 +1,15 @@
+import datetime
+
 import pytest
 from httpcore import Request, Response
 
 from hishel import AsyncFileStorage, AsyncRedisStorage
+from hishel._serializers import Metadata
 from hishel._utils import asleep, generate_key
+
+dummy_metadata = Metadata(
+    cache_key="test", number_of_uses=0, created_at=datetime.datetime.utcnow()
+)
 
 
 def is_redis_down() -> bool:
@@ -26,11 +33,13 @@ async def test_filestorage(use_temp_dir):
     response = Response(200, headers=[], content=b"test")
     await response.aread()
 
-    await storage.store(key, response=response, request=request)
+    await storage.store(
+        key, response=response, request=request, metadata=dummy_metadata
+    )
 
     stored_data = await storage.retreive(key)
     assert stored_data is not None
-    stored_response, stored_request = stored_data
+    stored_response, stored_request, metadata = stored_data
     stored_response.read()
     assert isinstance(stored_response, Response)
     assert stored_response.status == 200
@@ -50,11 +59,13 @@ async def test_redisstorage():
     response = Response(200, headers=[], content=b"test")
     await response.aread()
 
-    await storage.store(key, response=response, request=request)
+    await storage.store(
+        key, response=response, request=request, metadata=dummy_metadata
+    )
 
     stored_data = await storage.retreive(key)
     assert stored_data is not None
-    stored_response, stored_request = stored_data
+    stored_response, stored_request, metadata = stored_data
     stored_response.read()
     assert isinstance(stored_response, Response)
     assert stored_response.status == 200
@@ -74,10 +85,14 @@ async def test_filestorage_expired():
     response = Response(200, headers=[], content=b"test")
     await response.aread()
 
-    await storage.store(first_key, response=response, request=first_request)
+    await storage.store(
+        first_key, response=response, request=first_request, metadata=dummy_metadata
+    )
 
     await asleep(1)
-    await storage.store(second_key, response=response, request=second_request)
+    await storage.store(
+        second_key, response=response, request=second_request, metadata=dummy_metadata
+    )
 
     assert await storage.retreive(first_key) is None
 
@@ -95,9 +110,13 @@ async def test_redisstorage_expired():
     response = Response(200, headers=[], content=b"test")
     await response.aread()
 
-    await storage.store(first_key, response=response, request=first_request)
+    await storage.store(
+        first_key, response=response, request=first_request, metadata=dummy_metadata
+    )
 
     await asleep(1)
-    await storage.store(second_key, response=response, request=second_request)
+    await storage.store(
+        second_key, response=response, request=second_request, metadata=dummy_metadata
+    )
 
     assert await storage.retreive(first_key) is None